Added connectPrereqs function to util.js
